+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 1.6.0
 - Container customization can now be by Docker labels instead of, or in addition to, env vars
+- New container config to include or exclude image updates based on a regular expression
+### Internal
+- Many more test cases added, focusing on Docker and MQTT integration
 ## 1.5.2
 ### Internal
 - Build backend changed to `uv_build`